Prompt a yarn cache bust by upgrading prettier (#83)
The DocSearch deployment succeeded (as per Netlify logs) but I can't see
the search widget. I'm wondering whether this can be fixed by rebuilding
the application. The widget not showing up was something that happened a
couple of times locally for me, and busting the cache worked to fix it.