Rename structured logs to custom events in docs (#86)
Keeps the old page, which links to the new one (just until we have done the frontend release that links to the new page).

Will not merge until Bort 3.6.0 (with the renamed APIs) is released.

Still has a figma diagram referencing Structured Logs. I'll fix that up when I get access to it.