Update arm-mcuxpresso-guide.mdx (#230)
Need to update all "RT1060" to "i.MX RT1060" per a request from NXP Marketing that said:

"For the i.MX RT1060 Sample Project, if you do any social promotions, be sure to tag NXP and to use #NXPpartner.  Please be mindful not to abbreviate “i.MX RT1060”  (i.e. using just RT1060), whether on the Sample Project page or in promotions (in social media, you can switch to iMX without the dot). i.MX is a part of the RT brand."