Update nrf connect sdk guide for Bluetooth (#322)
* Update nrf connect sdk guide for Bluetooth

Now that we are a part of the Bluetooth SIG, we need to follow their rules of nomenclature. They do not allow "BLE". It needs to be Bluetooth Low Energy or Bluetooth LE.

Also Nordic updated our press release from all Bluetooth to Bluetooth LE, so we should update our docs to match their requests.